Rename `testnet` fixture to `ephnet`
As per the README update, changing the name of the fixture to `ephnet`
(short for 'ephemeral network') is intended to ensure a less
unambiguous name. The original name of `testnet` was chosen in
ignorance of the common use of the term to refer to a persistent
blockchain network used for testing.

Also changed:

 - 'existing' is used instead of 'persistent' in reference to
 ephemeral networks that are started and stopped with the `ephnetctl`
 tool. Referring to an ephemeral network as 'persistent' seemed
 confusing.

 - `TESTNETCTL_*`-prefixed env vars are now prefixed with
 `EPHNET_` (minus the `CTL`) since they are used by both the ephnetctl
 cli and the ginkgo test suite.

 - The root directory for storing local networks becomes `.ephnet`
 instead of `.testnetctl`, also due to the path being shared between
 the ephnetctl cli and the ginkgo test suite.